requirements/base.txt
aioredis==2.0.1
alembic==1.11.1
bcrypt
databases[asyncpg]==0.7.0
orjson==3.9.1
psycopg2==2.9.6
python-jose==3.3.0
httpx==0.24.1

pydantic[email]==1.10.9
fastapi==0.97.0
uvicorn[standard]==0.20.0

sentry-sdk==1.25.1
python-multipart==0.0.6
